Exercise 4.21 Blog deletion

+++ part4/exercises/backend/controllers/blogs.js
@@ -79,6 +79,19 @@ blogsRouter.post('/', async (request, response) => {
 
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  
+  const blog = await Blog.findById(request.params.id)
+  if (!blog.user) { // if the blog has no user attribute
+    return response.status(401).json({ error: 'This blog was created by another user' })
+  }
+  if (blog.user.toString() !== decodedToken.id) { // if the blog was created by another user
+    return response.status(401).json({ error: 'This blog was created by another user' })
+  }
+
   await Blog.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
